Adding Seller controller Get by id

+++ internal/domains/seller/SellerModel.go
@@ -1,9 +1,9 @@
 package seller
 
 type Seller struct {
-	Id          float64 `json:"id"`
-	Cid         float64 `json:"cid"`
-	CompanyName string  `json:"company_name"`
-	Address     string  `json:"address"`
-	Telephone   string  `json:"telephone"`
+	Id          int64  `json:"id"`
+	Cid         int64  `json:"cid"`
+	CompanyName string `json:"company_name"`
+	Address     string `json:"address"`
+	Telephone   string `json:"telephone"`
 }
