+++ internal/domains/seller/sellerService.go
@@ -2,7 +2,7 @@ package seller
 
 type Service interface {
 	GetAll() ([]Seller, error)
-	Get(id float64) (Seller, error)
+	Get(id int64) (Seller, error)
 }
 
 type service struct {
@@ -23,7 +23,7 @@ func (s *service) GetAll() ([]Seller, error) {
 	return listSeller, nil
 }
 
-func (s *service) Get(id float64) (Seller, error) {
+func (s *service) Get(id int64) (Seller, error) {
 	seller, err := s.repository.Get(id)
 
 	if err != nil {
